Update DESCRIPTION to add additional dependency packages.
Add dependencies to DESCRIPTION, most of which are needed for plotting.
Also remove 'collate', which appears to be depreciated (but can add that
back in if that was a mistake. Later commits will include additional R
functions for plotting and writing ghgvc data, so will use multiple
source files in the 'R/' directory.